Homework 3.A
Extend the existing application "Quiz". Create a new class "Quiz" that will represent the composition and/or aggregation of multiple questions & answers, complete the review and limit the addition of new questions to a maximum number of questions. The application must also allow adding a new question and answer via the keyboard. The application must print at the end the number of correct answers in absolute and percentage amount.