Homework 4.A
Define the base class "Person" that contains two attributes: first name and last name and corresponding getters and setters. Also create a method "displayDetails" which prints the details of a person.
Define a class "Student", which is inherited from the Person class, and contains additional attributes: student ID, field of enrolled subjects and teacher names. Within the class override the "displayDetails" method to print student details. 
Define a class "Teacher", which is inherited from the Person class. The class should have additional attributes: the names of the subjects held by the teacher and the salary amount. Override the "displayDetails" method to print teacher details.

Create "main" method in which at least two students and two teachers are instantiated, and call the "displayDetails" method for each of the objects.