Homework 8
Write a program that receives the names of two files via the command line. Each line of the file contains one name. The program should print all the names that are in the first file, but not in the second file.
The files must be relatively small (contain ten names)

the order in which the names are written is important.